Adds spec for when next step service. Moves next step service to workflow module.
